# frontend-react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bergen Smart Rain Hub API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/simulate/beredskap": {
      "post": {
        "summary": "Simulate Beredskap",
        "operationId": "simulate_beredskap_api_simulate_beredskap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeredskapsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeredskapsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BeredskapsRequest": {
        "properties": {
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/BuildingInput"
            },
            "type": "array",
            "title": "Buildings"
          },
          "tank_liters": {
            "type": "integer",
            "title": "Tank Liters"
          },
          "population": {
            "type": "integer",
            "title": "Population"
          },
          "efficiency": {
            "type": "number",
            "title": "Efficiency",
            "default": 0.85
          },
          "usage_level": {
            "type": "string",
            "title": "Usage Level",
            "default": "survival_total"
          },
          "climate_scenario": {
            "type": "string",
            "title": "Climate Scenario",
            "default": "historical"
          },
          "station": {
            "type": "string",
            "title": "Station",
            "default": "SN50540"
          }
        },
        "type": "object",
        "required": [
          "buildings",
          "tank_liters",
          "population"
        ],
        "title": "BeredskapsRequest"
      },
      "BeredskapsResponse": {
        "properties": {
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "simulation_series": {
            "items": {
              "$ref": "#/components/schemas/SimulationRow"
            },
            "type": "array",
            "title": "Simulation Series"
          },
          "dry_spells": {
            "items": {
              "$ref": "#/components/schemas/DrySpell"
            },
            "type": "array",
            "title": "Dry Spells"
          },
          "scenario_comparison": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ScenarioComparison"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Comparison"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "simulation_series",
          "dry_spells"
        ],
        "title": "BeredskapsResponse"
      },
      "BuildingInput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "roof_area_m2": {
            "type": "number",
            "title": "Roof Area M2"
          },
          "height_m": {
            "type": "number",
            "title": "Height M",
            "default": 5.0
          }
        },
        "type": "object",
        "required": [
          "name",
          "roof_area_m2"
        ],
        "title": "BuildingInput"
      },
...
      "DrySpell": {
        "properties": {
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "total_rain_mm": {
            "type": "number",
            "title": "Total Rain Mm"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "days",
          "total_rain_mm"
        ],
        "title": "DrySpell"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioComparison": {
        "properties": {
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "total_precip_mm": {
            "type": "number",
            "title": "Total Precip Mm"
          },
          "dry_days": {
            "type": "integer",
            "title": "Dry Days"
          },
          "longest_dry_spell": {
            "type": "integer",
            "title": "Longest Dry Spell"
          }
        },
        "type": "object",
        "required": [
          "scenario",
          "label",
          "total_precip_mm",
          "dry_days",
          "longest_dry_spell"
        ],
        "title": "ScenarioComparison"
      },
      "SimulationRow": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "precipitation_mm": {
            "type": "number",
            "title": "Precipitation Mm"
          },
          "tank_pct": {
            "type": "number",
            "title": "Tank Pct"
          },
          "tank_level_liters": {
            "type": "number",
            "title": "Tank Level Liters"
          },
          "days_remaining": {
            "type": "number",
            "title": "Days Remaining"
          }
        },
        "type": "object",
        "required": [
          "date",
          "precipitation_mm",
          "tank_pct",
          "tank_level_liters",
          "days_remaining"
        ],
        "title": "SimulationRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}